Move from a scheduled channel checking, to on-demand checking
As the old-channels channel is used infrequently enough, there is no
good justification for using a scheduled task for checking old
channels (unless the whole old-channels process from start to warning
to removal is automated).

This change means that the channel list is built on-demand when the
command is ran. A message is sent to first inform that the list is
being built; then, the message is edited afterwards with the true
list of inactive channels.